+++ src/main/kotlin/com/depromeet/makers/DepromeetMakersBeApplication.kt
@@ -1,6 +1,6 @@
 package com.depromeet.makers
 
-import com.depromeet.makers.domain.use_case.UseCase
+import com.depromeet.makers.domain.usecase.UseCase
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.context.annotation.ComponentScan
