+++ src/main/kotlin/com/depromeet/makers/domain/usecase/UseCase.kt
@@ -1,4 +1,4 @@
-package com.depromeet.makers.domain.use_case
+package com.depromeet.makers.domain.usecase
 
 interface UseCase<I, O> {
     fun execute(input: I): O
